gestion user, role admin

+++ backend/src/produits/produits.service.ts
@@ -1,9 +1,14 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  UnauthorizedException,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Produit } from './entities/produit.entity';
 import { CreateProduitDto } from './dto/create-produit.dto';
 import { UpdateProduitDto } from './dto/update-produit.dto';
+import { UserPayload } from 'src/auth/interfaces/user-payload.interface';
 
 @Injectable()
 export class ProduitsService {
@@ -12,7 +17,15 @@ export class ProduitsService {
     private readonly produitRepository: Repository<Produit>,
   ) {}
 
-  async create(createProduitDto: CreateProduitDto): Promise<Produit> {
+  async create(
+    createProduitDto: CreateProduitDto,
+    user: UserPayload,
+  ): Promise<Produit> {
+    if (user.role !== 'admin') {
+      throw new UnauthorizedException(
+        'Seuls les administrateurs peuvent créer des produits.',
+      );
+    }
     const produit = this.produitRepository.create(createProduitDto);
     return await this.produitRepository.save(produit);
   }
@@ -32,7 +45,13 @@ export class ProduitsService {
   async update(
     id: number,
     updateProduitDto: UpdateProduitDto,
+    user: UserPayload,
   ): Promise<Produit> {
+    if (user.role !== 'admin') {
+      throw new UnauthorizedException(
+        'Seuls les administrateurs peuvent mettre à jour des produits.',
+      );
+    }
     const produit = await this.produitRepository.preload({
       id: id,
       ...updateProduitDto,
@@ -40,13 +59,25 @@ export class ProduitsService {
     if (!produit) {
       throw new NotFoundException(`Produit #${id} introuvable`);
     }
-    return this.produitRepository.save(produit);
+    return await this.produitRepository.save(produit);
   }
 
-  async remove(id: number): Promise<void> {
+  async remove(id: number, user: UserPayload): Promise<void> {
+    if (user.role !== 'admin') {
+      throw new UnauthorizedException(
+        'Seuls les administrateurs peuvent supprimer des produits.',
+      );
+    }
     const result = await this.produitRepository.delete(id);
     if (result.affected === 0) {
       throw new NotFoundException(`Produit #${id} introuvable`);
     }
+  }
+
+  // Nouvelle méthode pour récupérer les produits par catégorie
+  async findByCategory(categoryId: number): Promise<Produit[]> {
+    return await this.produitRepository.find({
+      where: { categorie_id: categoryId },
+    });
   }
 }
